move spells into api folder on server
