+++ packages/server-core/src/index.ts
@@ -15,9 +15,8 @@ export * from './servers/queryGoogleSearch'
 export * from './utils/ServerError'
 export * from './services'
 export * from './sockets/sockets'
-export * from './spells'
+export * from './api'
 export * from './servers/tiktalknet'
 export * from './config/types'
 export * from './utils'
 export * from './config/validators'
-
